JWT auth and admin authorization for protected routes

+++ routes/campsiteRouter.js
@@ -1,26 +1,28 @@
 const express = require("express");
-const campsiteRouter = express.Router();
 const Campsite = require("../models/campsite");
 const authenticate = require("../authenticate");
+
+const campsiteRouter = express.Router();
 
 campsiteRouter
   .route("/")
   .get((req, res, next) => {
+    // req and res are objects being passed into the function
     Campsite.find()
       .populate("comments.author")
       .then((campsites) => {
         res.statusCode = 200;
-        res.setHeader("content-type", "application/json");
+        res.setHeader("Content-Type", "application/json");
         res.json(campsites);
       })
       .catch((err) => next(err));
   })
-  .post(authenticate.verifyUser, (req, res, next) => {
+  .post(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
     Campsite.create(req.body)
       .then((campsite) => {
-        console.log("campsite created: ", campsite);
+        console.log("Campsite Created ", campsite);
         res.statusCode = 200;
-        res.setHeader("content-type", "application/json");
+        res.setHeader("Content-Type", "application/json");
         res.json(campsite);
       })
       .catch((err) => next(err));
@@ -29,15 +31,19 @@ campsiteRouter
     res.statusCode = 403;
     res.end("PUT operation not supported on /campsites");
   })
-  .delete(authenticate.verifyUser, (req, res, next) => {
-    Campsite.deleteMany()
-      .then((response) => {
-        res.statusCode = 200;
-        res.setHeader("content-type", "application/json");
-        res.json(response);
-      })
-      .catch((err) => next(err));
-  });
+  .delete(
+    authenticate.verifyUser,
+    authenticate.verifyAdmin,
+    (req, res, next) => {
+      Campsite.deleteMany()
+        .then((response) => {
+          res.statusCode = 200;
+          res.setHeader("Content-Type", "application/json");
+          res.json(response);
+        })
+        .catch((err) => next(err));
+    }
+  );
 
 campsiteRouter
   .route("/:campsiteId")
@@ -46,47 +52,56 @@ campsiteRouter
       .populate("comments.author")
       .then((campsite) => {
         res.statusCode = 200;
-        res.setHeader("content-type", "application/json");
+        res.setHeader("Content-Type", "application/json");
         res.json(campsite);
       })
       .catch((err) => next(err));
   })
   .post(authenticate.verifyUser, (req, res) => {
     res.statusCode = 403;
-    res.end(`POST operation not support on campsites/${req.params.campsiteId}`);
+    res.end(
+      `POST operation not supported on /campsites/${req.params.campsiteId}`
+    );
   })
-  .put(authenticate.verifyUser, (req, res, next) => {
+  .put(authenticate.verifyUser, authenticate.verifyAdmin, (req, res, next) => {
     Campsite.findByIdAndUpdate(
       req.params.campsiteId,
-      { $set: req.body },
+      {
+        $set: req.body,
+      },
       { new: true }
     )
       .then((campsite) => {
         res.statusCode = 200;
-        res.setHeader("content-type", "application/json");
+        res.setHeader("Content-Type", "application/json");
         res.json(campsite);
       })
       .catch((err) => next(err));
   })
-  .delete(authenticate.verifyUser, (req, res, next) => {
-    Campsite.findByIdAndDelete(req.params.campsiteId)
-      .then((campsite) => {
-        res.statusCode = 200;
-        res.setHeader("content-type", "application/json");
-        res.json(campsite);
-      })
-      .catch((err) => next(err));
-  });
+  .delete(
+    authenticate.verifyUser,
+    authenticate.verifyAdmin,
+    (req, res, next) => {
+      Campsite.findByIdAndDelete(req.params.campsiteId)
+        .then((response) => {
+          res.statusCode = 200;
+          res.setHeader("Content-Type", "application/json");
+          res.json(response);
+        })
+        .catch((err) => next(err));
+    }
+  );
 
 campsiteRouter
-  .route("/:campsiteId/comments") //ALL COMMENTS
+  .route("/:campsiteId/comments")
   .get((req, res, next) => {
+    // req and res are objects being passed into the function
     Campsite.findById(req.params.campsiteId)
       .populate("comments.author")
       .then((campsite) => {
         if (campsite) {
           res.statusCode = 200;
-          res.setHeader("content-type", "application/json");
+          res.setHeader("Content-Type", "application/json");
           res.json(campsite.comments);
         } else {
           err = new Error(`Campsite ${req.params.campsiteId} not found`);
@@ -96,7 +111,7 @@ campsiteRouter
       })
       .catch((err) => next(err));
   })
-  .post(authenticate.verifyUser, (req, res) => {
+  .post(authenticate.verifyUser, (req, res, next) => {
     Campsite.findById(req.params.campsiteId)
       .then((campsite) => {
         if (campsite) {
@@ -106,7 +121,7 @@ campsiteRouter
             .save()
             .then((campsite) => {
               res.statusCode = 200;
-              res.setHeader("content-type", "application/json");
+              res.setHeader("Content-Type", "application/json");
               res.json(campsite);
             })
             .catch((err) => next(err));
@@ -121,42 +136,47 @@ campsiteRouter
   .put(authenticate.verifyUser, (req, res) => {
     res.statusCode = 403;
     res.end(
-      `PUT operations not support on campsites/${req.params.campsiteId}/comments`
+      `PUT operation not supported on /campsites/${req.params.campsiteId}/comments`
     );
   })
-  .delete(authenticate.verifyUser, (req, res, next) => {
-    Campsite.findById(req.params.campsiteId)
-      .then((campsite) => {
-        if (campsite) {
-          for (let i = campsite.comments.length - 1; i >= 0; i--) {
-            campsite.comments.id(campsite.comments[i]._id).deleteOne();
+  .delete(
+    authenticate.verifyUser,
+    authenticate.verifyAdmin,
+    (req, res, next) => {
+      Campsite.findById(req.params.campsiteId)
+        .then((campsite) => {
+          if (campsite) {
+            for (let i = campsite.comments.length - 1; i >= 0; i--) {
+              campsite.comments.id(campsite.comments[i]._id).deleteOne();
+            }
+            campsite
+              .save()
+              .then((campsite) => {
+                res.statusCode = 200;
+                res.setHeader("Content-Type", "application/json");
+                res.json(campsite);
+              })
+              .catch((err) => next(err));
+          } else {
+            err = new Error(`Campsite ${req.params.campsiteId} not found`);
+            err.status = 404;
+            return next(err);
           }
-          campsite
-            .save()
-            .then((campsite) => {
-              res.statusCode = 200;
-              res.setHeader("content-type", "application/json");
-              res.json(campsite);
-            })
-            .catch((err) => next(err));
-        } else {
-          err = new Error(`Campsite ${req.params.campsiteId} not found`);
-          err.status = 404;
-          return next(err);
-        }
-      })
-      .catch((err) => next(err));
-  });
+        })
+        .catch((err) => next(err));
+    }
+  );
 
 campsiteRouter
-  .route("/:campsiteId/comments/:commentId") //ALL COMMENTS
+  .route("/:campsiteId/comments/:commentId")
   .get((req, res, next) => {
+    // req and res are objects being passed into the function
     Campsite.findById(req.params.campsiteId)
       .populate("comments.author")
       .then((campsite) => {
         if (campsite && campsite.comments.id(req.params.commentId)) {
           res.statusCode = 200;
-          res.setHeader("content-type", "application/json");
+          res.setHeader("Content-Type", "application/json");
           res.json(campsite.comments.id(req.params.commentId));
         } else if (!campsite) {
           err = new Error(`Campsite ${req.params.campsiteId} not found`);
@@ -173,13 +193,18 @@ campsiteRouter
   .post(authenticate.verifyUser, (req, res) => {
     res.statusCode = 403;
     res.end(
-      `POST operations not support on /campsites/${req.params.campsiteId}/comments/${req.params.commentId}`
+      `POST operation not supported on /campsites/${req.params.campsiteId}/comments/${req.params.commentId}`
     );
   })
   .put(authenticate.verifyUser, (req, res, next) => {
     Campsite.findById(req.params.campsiteId)
       .then((campsite) => {
         if (campsite && campsite.comments.id(req.params.commentId)) {
+          const comment = campsite.comments.id(req.params.commentId);
+          if (!comment.author.equals(req.user._id)) {
+            res.statusCode = 403;
+            return res.json({ err: "You are not authorized" });
+          }
           if (req.body.rating) {
             campsite.comments.id(req.params.commentId).rating = req.body.rating;
           }
@@ -190,12 +215,12 @@ campsiteRouter
             .save()
             .then((campsite) => {
               res.statusCode = 200;
-              res.setHeader("content-type", "application/json");
+              res.setHeader("Content-Type", "application/json");
               res.json(campsite);
             })
             .catch((err) => next(err));
         } else if (!campsite) {
-          err = new Error(`Comment ${req.params.commentId} not found`);
+          err = new Error(`Campsite ${req.params.campsiteId} not found`);
           err.status = 404;
           return next(err);
         } else {
@@ -206,22 +231,26 @@ campsiteRouter
       })
       .catch((err) => next(err));
   })
-
   .delete(authenticate.verifyUser, (req, res, next) => {
     Campsite.findById(req.params.campsiteId)
       .then((campsite) => {
         if (campsite && campsite.comments.id(req.params.commentId)) {
+          const comment = campsite.comments.id(req.params.commentId);
+          if (!comment.author.equals(req.user._id)) {
+            res.statusCode = 403;
+            return res.json({ err: "You are not authorized" });
+          }
           campsite.comments.id(req.params.commentId).deleteOne();
           campsite
             .save()
             .then((campsite) => {
               res.statusCode = 200;
-              res.setHeader("content-type", "application/json");
+              res.setHeader("Content-Type", "application/json");
               res.json(campsite);
             })
             .catch((err) => next(err));
         } else if (!campsite) {
-          err = new Error(`Comment ${req.params.commentId} not found`);
+          err = new Error(`Campsite ${req.params.campsiteId} not found`);
           err.status = 404;
           return next(err);
         } else {
